Separate tests into separate files based on which route is being tested Add test for /charge route with a correct payload (i.e. a correct token) Relates #34 #36
