+++ tests/src/routes/home.js
@@ -1,0 +1,14 @@
+const test = require('tape');
+const server = require('../../../src/server.js');
+
+test('test that the home route returns a status code of 200', (t) => {
+  const options = {
+    method: 'GET',
+    url: '/',
+  };
+
+  server.inject(options, (res) => {
+    t.equal(res.statusCode, 200, 'should return status code of 200');
+    t.end();
+  });
+});
